# server/openapi.yaml
openapi: 3.1.0
info:
  title: Stars Trail API
  version: v1
  description: |
    Contract between the Android app (Kotlin) and the Go server (spec §3.2); the two share no code.
    Every path is under /v1. The app sends on every request:
    - `X-Device-Id`: an anonymous per-install ID, the key for the weather proxy's daily quota (falls back to the caller's IP).
    - `X-Client-Version`: its Android versionCode. Below `minClientVersion` every route except
      /v1/health and /v1/version answers 426; offline features never depend on the server.
      A request without the header is not gated (curl, monitoring), so the app must always send it:
      an app that forgets it can never be forced to update.
    Every future feature path lists both headers and the 426 response.
    Routes marked with bearerAuth need `Authorization: Bearer <token>` from /v1/auth/login; a missing,
    unknown or logged-out token is 401 unauthorized. Everything else is anonymous.
    Every error body is an Error: a code from ErrorCode plus the fields that code documents (ADR 0004).
    The Go server is generated from this file (`go generate ./...`, oapi-codegen): edit here first.
servers:
  - url: /v1
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, description: 'the token from /v1/auth/login; valid until logout'}
  parameters:
    DeviceId: {name: X-Device-Id, in: header, schema: {type: string}}
    ClientVersion: {name: X-Client-Version, in: header, schema: {type: integer}}
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - invalid_request      # body or parameters don't parse
        - internal             # server-side failure; details only in the server log
        - rate_limited         # login tries only
        - client_outdated
        - invalid_region
        - region_too_large
        - region_unsupported
        - daily_quota_exceeded
        - data_unavailable
        - unauthorized         # no, unknown or logged-out bearer token
        - invalid_phone        # not an 11-digit +86 mobile number
        - wrong_code           # the SMS code doesn't match, or has expired
        - sms_too_frequent     # this number was sent too many codes; wait before asking again
        - sms_unavailable      # the SMS provider failed or isn't configured
        - team_not_found       # no active team has this code, or the caller isn't a member of this team
        - not_initiator        # only the 发起人 may end the trip or set the 队伍轨迹
        - team_ended           # the trip was ended: no more positions or sharing
        - image_not_found      # no such image in this team, or no such synced photo of the caller's
        - photo_quota_exceeded # the caller's synced photos would pass 1 GB (spec §2.12); keep this one on the phone
        - no_team_track        # the team has no 队伍轨迹 (none given, 取消, or the trip ended)
    Error:
      type: object
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorCode'}
        minClientVersion: {type: integer, description: set when error is client_outdated}
        maxAreaKm2: {type: integer, description: set when error is region_too_large}
        quotaBytes: {type: integer, format: int64, description: set when error is photo_quota_exceeded}
        quotaCells: {type: integer, format: int64, description: set when error is daily_quota_exceeded on weather}
...
    TeamRequest:
      type: object
      description: Nothing to give; members are shown by their account's nickname (a name sent by older apps is ignored).
...
    Team:
      type: object
      description: |
        A 队伍 as the caller sees it. Positions and messages are those stored after the `after` cursor the
        request gave (all of them without one); `cursor` is the last stored so far, to pass as `after` next time.
      required: [id, code, initiator, me, ended, cursor, members, messages]
      properties:
        id: {type: integer, format: int64}
        me: {type: integer, format: int64, description: the caller's user id, to find themselves among the members}
        code: {type: string, description: 4-digit join code; free for another team once the trip ends}
        initiator: {type: integer, format: int64, description: user id of the 发起人}
        ended: {type: boolean, description: 结束行程 happened; nobody shares any more}
        cursor: {type: integer, format: int64}
        members: {type: array, items: {$ref: '#/components/schemas/Member'}}
        messages: {type: array, items: {$ref: '#/components/schemas/Message'}, description: '队伍对话, in the order stored'}
        track: {$ref: '#/components/schemas/TeamTrackRef'}
    TeamTrackRef:
      type: object
      description: 'The 队伍轨迹 (spec §2.11) without its points, which GET /teams/{id}/track has; absent when there is none.'
      required: [version, uuid, name, reversed, start]
      properties:
        version: {type: integer, format: int64, description: 'grows with every 指定, 更换, 改起算点 and 取消: a member fetches the points when it changes'}
        uuid: {type: string, maxLength: 64, description: 'the 发起人''s track it was taken from: the same track goes into a member''s 我的轨迹 once'}
        name: {type: string, minLength: 1, maxLength: 100}
        reversed: {type: boolean, description: '起算点: 反向'}
        start: {type: number, format: double, minimum: 0, description: 'a loop''s 起点, metres along the points as they go (0: their own start)'}
...
    Member:
      type: object
      required: [id, name, sharing, positions]
      properties:
        id: {type: integer, format: int64, description: user id}
        name: {type: string, description: their account's nickname as it is now}
        sharing: {type: boolean, description: false after 停止共享 (or 结束行程)}
        positions: {type: array, items: {$ref: '#/components/schemas/Position'}, description: 'in the order stored; 尾迹 sorts by time'}
    Position:
      type: object
      required: [time, lat, lon]
      properties:
        time: {type: integer, format: int64, description: 'fix time, Unix seconds'}
        lat: {type: number, format: double, minimum: -90, maximum: 90}
        lon: {type: number, format: double, minimum: -180, maximum: 180}
        battery: {type: integer, minimum: 0, maximum: 100, description: percent}
...
    MessageKind:
      type: string
      enum:
        - text       # text
        - location   # lat, lon, along
        - image      # image, from /teams/{id}/images
        - system     # text: the server's own note (队伍轨迹 given, changed or dropped), from the member it's about; not sendable
...
    Message:
      type: object
      description: A MessageRequest as stored, with who sent it and when.
      required: [seq, name, time, kind]
      properties:
        seq: {type: integer, format: int64, description: 'order within the team, comparable with a Team''s cursor: unread are those after the last one read'}
        from: {type: integer, format: int64, description: 'sender''s user id; absent once their account is deleted'}
        name: {type: string, description: 'the sender''s nickname as it is now (they may have left since); 已注销用户 once their account is deleted'}
        time: {type: integer, format: int64, description: 'server time, Unix seconds'}
        kind: {$ref: '#/components/schemas/MessageKind'}
        text: {type: string}
        lat: {type: number, format: double}
        lon: {type: number, format: double}
        image: {type: string}
        along: {type: array, items: {type: number, format: double}}
...
    SyncPoint:
      type: object
      required: [t, lat, lon, s]
      properties:
        t: {type: integer, format: int64, description: 'fix time, Unix milliseconds; 0 when unknown (a planned or time-less import)'}
        lat: {type: number, format: double, description: as recorded or imported; the track's datum says how to read it}
        lon: {type: number, format: double}
        ele: {type: number, format: double, description: metres; absent if unknown}
        s: {type: integer, description: segment (a track is paused and continued in segments)}
    SyncTrack:
      type: object
      description: |
        A 轨迹 as synced (spec §2.12), keyed by the id the phone made for it. Only ended tracks sync, and their
        points never change; the attributes (name, datum) do. deleted is a 删除标记: the track is gone, its
        points and attributes with it. public: a 公开轨迹 (spec §2.8), shown to everyone in /tiles/public-tracks.
      required: [id, startedAt, endedAt, planned, points, name, datum, public, deleted]
      properties:
        id: {type: string, description: 32 lowercase hex digits}
        startedAt: {type: integer, format: int64, description: Unix milliseconds}
        endedAt: {type: integer, format: int64, description: Unix milliseconds}
        planned: {type: boolean, description: 计划轨迹}
        points: {type: array, items: {$ref: '#/components/schemas/SyncPoint'}, description: in the order recorded or imported}
        name: {type: string, description: empty for a recording (shown by its start time)}
        datum: {$ref: '#/components/schemas/Datum'}
        public: {type: boolean}
        deleted: {type: boolean}
    SyncWaypoint:
      type: object
      description: A 标注 as synced. Its place and time never change; name, description, photo and group do. deleted as in SyncTrack.
      required: [id, track, group, time, lat, lon, name, description, photo, deleted]
      properties:
        id: {type: string, description: 32 lowercase hex digits}
        track: {type: string, description: 'the SyncTrack it was added on, or empty'}
        group: {type: string, description: 'the SyncGroup it is in, or empty; never set with track'}
        time: {type: integer, format: int64, description: 'Unix milliseconds; 0 when unknown'}
        lat: {type: number, format: double}
        lon: {type: number, format: double}
        ele: {type: number, format: double}
        name: {type: string}
        description: {type: string}
        photo: {type: string, description: 'a photo from /sync/photos, or empty'}
        deleted: {type: boolean}
    SyncGroup:
      type: object
      description: |
        A 标注组 as synced (#121), keyed by the id the phone made for it. Names are unique per account on the
        phones, not here: a phone that pulls a name it already has renames its copy and pushes that back.
        Deleting a group deletes its 标注: the phone pushes their 删除标记 too. deleted as in SyncTrack.
      required: [id, name, deleted]
      properties:
        id: {type: string, description: 32 lowercase hex digits}
        name: {type: string}
        deleted: {type: boolean}
    Datum:
      type: string
      enum: [WGS84, GCJ02, BD09]
      description: 坐标纠偏 the track's points (and its 标注) are read with
    SyncTrackChange:
      type: object
      description: |
        What changed of a track. A track new to the server needs startedAt, endedAt, planned and points (else
        invalid_request); on a known one those are ignored. Attributes left out stay as they are; those sent
        overwrite, each on its own: the last the server receives wins. After deleted, changes are ignored.
      required: [id]
      properties:
        id: {type: string}
        startedAt: {type: integer, format: int64}
        endedAt: {type: integer, format: int64}
        planned: {type: boolean}
        points: {type: array, items: {$ref: '#/components/schemas/SyncPoint'}}
        name: {type: string, maxLength: 200}
        datum: {$ref: '#/components/schemas/Datum'}
        public: {type: boolean, description: 'publish as a 公开轨迹, or withdraw it; deleting withdraws it too'}
        deleted: {type: boolean, description: 'true: 删除标记; false is ignored'}
    SyncWaypointChange:
      type: object
      description: As SyncTrackChange; a new 标注 needs time, lat and lon. photo must be one of the caller's, or empty.
      required: [id]
      properties:
        id: {type: string}
        track: {type: string}
        group: {type: string, description: 'a SyncGroup id, or empty for none; overwrites like an attribute'}
        time: {type: integer, format: int64}
        lat: {type: number, format: double}
        lon: {type: number, format: double}
        ele: {type: number, format: double}
        name: {type: string, maxLength: 200}
        description: {type: string, maxLength: 10000}
        photo: {type: string}
        deleted: {type: boolean, description: 'true: 删除标记; false is ignored'}
    SyncGroupChange:
      type: object
      description: As SyncTrackChange; a new group needs name.
      required: [id]
      properties:
        id: {type: string}
        name: {type: string, maxLength: 200}
        deleted: {type: boolean, description: 'true: 删除标记; false is ignored'}
    SyncChanges:
      type: object
      required: [tracks, waypoints]
      properties:
        tracks: {type: array, items: {$ref: '#/components/schemas/SyncTrackChange'}}
        groups: {type: array, items: {$ref: '#/components/schemas/SyncGroupChange'}, description: 'stored before the waypoints'}
        waypoints: {type: array, items: {$ref: '#/components/schemas/SyncWaypointChange'}}
    Sync:
      type: object
      description: 'Tracks, 标注组 and 标注 changed after the `after` cursor (by anyone: the caller''s own writes come back too), in the order changed.'
      required: [cursor, more, tracks, groups, waypoints]
      properties:
        cursor: {type: integer, format: int64, description: pass as after next time}
        more: {type: boolean, description: 'there are more changes: ask again with the cursor'}
        tracks: {type: array, items: {$ref: '#/components/schemas/SyncTrack'}}
        groups: {type: array, items: {$ref: '#/components/schemas/SyncGroup'}}
        waypoints: {type: array, items: {$ref: '#/components/schemas/SyncWaypoint'}}
...
  responses:
...
    ClientOutdated:
      description: Client older than minClientVersion (client_outdated); the app shows "update required" on this online feature.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Internal:
      description: Server-side failure (internal).
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Unauthorized:
      description: No, unknown or logged-out bearer token (unauthorized); the app logs out locally.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
paths:
...
  /sync:
    get:
      operationId: getSync
      summary: Pull the caller's synced tracks, 标注组 and 标注 changed after a cursor (spec §2.12)
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/DeviceId'}
        - {$ref: '#/components/parameters/ClientVersion'}
        - {name: after, in: query, schema: {type: integer, format: int64}, description: 'a Sync''s cursor; none for everything'}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Sync'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
    post:
      operationId: postSync
      summary: Push changes to the caller's tracks, 标注组 and 标注, all or nothing
      description: 'The body may be up to 32 MB (a long track''s points); everything else stays at 1 MB.'
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SyncChanges'}}}
      responses:
        '204': {description: Stored}
        '400': {description: 'invalid_request (a bad id, or a new one missing what it needs), or image_not_found', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
...
  /teams:
    post:
      operationId: postTeam
      summary: Create a 队伍 (spec §2.11); the caller is its 发起人 and gets a 4-digit code to hand out
      description: A caller already in another active team leaves it first. The caller starts out sharing.
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/DeviceId'}, {$ref: '#/components/parameters/ClientVersion'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/TeamRequest'}}}
      responses:
        '200': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Team'}}}}
        '400': {description: invalid_request, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '426': {$ref: '#/components/responses/ClientOutdated'}
        '500': {$ref: '#/components/responses/Internal'}
...